fix(nrgu): sell with decimal quantities

+++ nrgu_live.py
@@ -105,7 +105,7 @@ def loop(symbol: str):
             buy_symbol_market(symbol, target_quantity)
 
         elif position and sell_reason:
-            position_quantity = int(position["qty"])
+            position_quantity = float(position["qty"])
             print(f"selling, {rsi=} {williamsr=} {position_quantity=}")
             sell_symbol_market(symbol, position_quantity)
